+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.39.0] - 2026-03-15
+[v1.39.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.39.0
+
 ### Added
 
 * New tap action: remove photo